@@ -16,7 +16,9 @@
   ],
   "repository": "LinusU/ts-unreachable",
   "license": "MIT",
+  "type": "module",
+  "exports": "./index.js",
   "engines": {
-    "node": ">=8.10"
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   }
 }
