@@ -1,3 +1,3 @@
-module.exports = function unreachable (value) {
+export default function unreachable (value) {
   throw new TypeError('Encountered impossible value: ' + JSON.stringify(value))
 }
